python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-gateway"
version = "0.1.0"
description = "AI gateway for managing and routing LLM requests - Govern, Secure, and Optimize your AI Traffic"
authors = [
    {name = "AI Gateway Team", email = "api@example.com"},
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "openai>=1.13.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0", # For Google Gemini
    "boto3>=1.34.0",  # For AWS Bedrock
    "tiktoken>=0.6.0", # Token counting for various models
    "azure-identity>=1.15.0", # For Azure OpenAI
    "azure-openai>=1.0.0", # For Azure OpenAI
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp>=1.23.0",
    "pyyaml>=6.0.1",
    "redis>=5.0.1", # For caching
    "click>=8.1.7", # For CLI
    "rich>=13.7.0", # For TUI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.urls]
"Homepage" = "https://github.com/example/ai-gateway"
"Bug Tracker" = "https://github.com/example/ai-gateway/issues"

[project.scripts]
ai-gateway = "gateway.cli.commands:main"

[tool.setuptools]
package-dir = {}

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true